[NEW FEATURES] Added Slider and Basic CMS Blog
I have added a basic slider which can be extended past 4 with ease,
needs styling and images adding but this is mainly all for testing

I have also added a basic cms blog system with adding/editing/deleting
functionailty. This is intended to make your life easier when posting
updates.

Also added a view page so people can view the entire update as the
homepage only shows the title and the date it was posted

Sweet! Thats all for now!